Added "." Changed type of operands from Int to Double Processing divide on zero

+++ app/src/main/java/com/example/calculatorapp/Operation.kt
@@ -2,27 +2,27 @@ package com.example.calculatorapp
 
 enum class Operation(val sign: String) {
     PLUS("+") {
-        override fun performOperation(first: Int, second: Int): Int {
+        override fun performOperation(first: Float, second: Float): Float {
             return first + second
         }
     },
     MINUS("-") {
-        override fun performOperation(first: Int, second: Int): Int {
+        override fun performOperation(first: Float, second: Float): Float {
             return first - second
         }
     },
     DIVIDE("/") {
-        override fun performOperation(first: Int, second: Int): Int {
+        override fun performOperation(first: Float, second: Float): Float {
             return first / second
         }
     },
     MULTIPLY("*") {
-        override fun performOperation(first: Int, second: Int): Int {
+        override fun performOperation(first: Float, second: Float): Float {
             return first * second
         }
     };
 
-    abstract fun performOperation(first: Int, second: Int): Int
+    abstract fun performOperation(first: Float, second: Float): Float
 
     companion object{
         fun value(sign: String): Operation? {
